# docs/API_SPEC_OSHC_Platform.yaml
openapi: 3.1.0
info:
  title: OSHC Comparison Platform API
  version: 0.1.0
  summary: MVP API contract for the OSHC comparison, checkout, admin, agent, and fulfilment portal.
  description: >
    Swagger/OpenAPI-compatible API specification for the OSHC comparison and purchase platform.
    This contract is intended for frontend/backend alignment, Swagger UI review, test generation,
    and implementation planning.
servers:
  - url: https://api.example.com/v1
    description: Production placeholder
  - url: https://staging-api.example.com/v1
    description: Staging placeholder
  - url: http://localhost:3000/api/v1
    description: Local development placeholder
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    post:
      tags: [Checkout]
      summary: Create order from application
      security: []
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /orders/{orderId}:
    get:
      tags: [Checkout]
      summary: Get public order summary
      description: Returns safe order summary for confirmation pages. Customer self-service lookup is post-MVP.
      security: []
      operationId: getOrder
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/auth/login:
    post:
      tags: [Admin Auth]
      summary: Admin login
      security: []
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/providers:
    get:
      tags: [Admin Providers]
      summary: List providers
      operationId: adminListProviders
      responses:
        "200":
          description: Provider list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Provider"
    post:
      tags: [Admin Providers]
      summary: Create provider
      operationId: adminCreateProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProviderRequest"
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /admin/providers/{providerId}:
    patch:
      tags: [Admin Providers]
      summary: Update provider
      operationId: adminUpdateProvider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProviderRequest"
      responses:
        "200":
          description: Provider updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/products:
    get:
      tags: [Admin Products]
      summary: List products
      operationId: adminListProducts
      parameters:
        - name: providerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
    post:
      tags: [Admin Products]
      summary: Create product
      operationId: adminCreateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /admin/products/{productId}:
    patch:
      tags: [Admin Products]
      summary: Update product
      operationId: adminUpdateProduct
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /admin/price-imports:
    post:
      tags: [Admin Prices]
      summary: Upload price table CSV
      operationId: adminUploadPriceImport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [providerId, file]
              properties:
                providerId:
                  type: string
                  format: uuid
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Import uploaded and validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceImportResponse"
  /admin/price-imports/{priceImportId}/publish:
    post:
      tags: [Admin Prices]
      summary: Publish validated price import
      operationId: adminPublishPriceImport
      parameters:
        - name: priceImportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Price import published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceImportResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /admin/orders:
    get:
      tags: [Admin Orders]
      summary: Search admin orders
      operationId: adminSearchOrders
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: providerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Order search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderListResponse"
  /admin/orders/{orderId}:
    get:
      tags: [Admin Orders]
      summary: Get admin order detail
      operationId: adminGetOrder
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Admin order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetailResponse"
  /admin/orders/{orderId}/certificates:
    post:
      tags: [Fulfilment]
      summary: Upload order certificate
      operationId: adminUploadCertificate
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [policyNumber, file]
              properties:
                policyNumber:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Certificate uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateResponse"
  /admin/fulfilment-queue:
    get:
      tags: [Fulfilment]
      summary: List orders awaiting fulfilment
      operationId: adminGetFulfilmentQueue
      parameters:
        - name: providerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: breachedOnly
          in: query
          required: false
          schema:
            type: boolean
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Fulfilment queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FulfilmentQueueResponse"
...
  /admin/reports/quote-funnel:
    get:
      tags: [Reports]
      summary: Quote funnel report
      operationId: adminQuoteFunnelReport
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
      responses:
        "200":
          description: Quote funnel metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteFunnelReportResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    FromDate:
      name: from
      in: query
      required: true
      schema:
        type: string
        format: date
    ToDate:
      name: to
      in: query
      required: true
      schema:
        type: string
        format: date
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Resource not found.
            requestId:
              type: string
...
    OrderStatus:
      type: string
      enum:
        - quote_created
        - application_started
        - application_completed
        - payment_pending
        - payment_failed
        - paid
        - provider_submission_pending
        - provider_submission_failed
        - certificate_pending
        - certificate_issued
        - fulfilled
        - refund_requested
        - refunded
        - cancelled
...
    CreateOrderRequest:
      type: object
      required: [applicationId]
      properties:
        applicationId:
          type: string
          format: uuid
    Order:
      type: object
      required: [id, orderReference, status, total, customerEmail]
      properties:
        id:
          type: string
          format: uuid
        orderReference:
          type: string
          example: OSHC-20260601-000123
        status:
          $ref: "#/components/schemas/OrderStatus"
        total:
          $ref: "#/components/schemas/Money"
        customerEmail:
          type: string
          format: email
        providerName:
          type: string
        productName:
          type: string
        paidAt:
          type: string
          format: date-time
          nullable: true
    OrderResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Order"
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [accessToken, tokenType, expiresIn]
          properties:
            accessToken:
              type: string
            tokenType:
              type: string
              example: Bearer
            expiresIn:
              type: integer
              example: 3600
    Provider:
      type: object
      required: [id, code, name, status]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/ProviderStatus"
        logoUrl:
          type: string
          format: uri
          nullable: true
        disclosureUrl:
          type: string
          format: uri
          nullable: true
        productDisclosureUrl:
          type: string
          format: uri
          nullable: true
    UpsertProviderRequest:
      type: object
      required: [code, name, status]
      properties:
        code:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/ProviderStatus"
        logoFileRef:
          type: string
        disclosureUrl:
          type: string
          format: uri
        productDisclosureUrl:
          type: string
          format: uri
    ProviderResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Provider"
    Product:
      type: object
      required: [id, providerId, code, name, status, coverTypes]
      properties:
        id:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/ProductStatus"
        coverTypes:
          type: array
          items:
            $ref: "#/components/schemas/CoverType"
        disclosureUrl:
          type: string
          format: uri
          nullable: true
        policyDocumentUrl:
          type: string
          format: uri
          nullable: true
    UpsertProductRequest:
      type: object
      required: [providerId, code, name, status, coverTypes]
      properties:
        providerId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/ProductStatus"
        coverTypes:
          type: array
          items:
            $ref: "#/components/schemas/CoverType"
        disclosureUrl:
          type: string
          format: uri
        policyDocumentUrl:
          type: string
          format: uri
    ProductResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Product"
    PriceImportResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, status, validationSummary]
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum: [draft, validated, published, rejected, archived]
            validationSummary:
              type: object
              properties:
                totalRows:
                  type: integer
                validRows:
                  type: integer
                invalidRows:
                  type: integer
                errors:
                  type: array
                  items:
                    type: object
                    properties:
                      row:
                        type: integer
                      field:
                        type: string
                      message:
                        type: string
    AdminOrderListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrderSummary"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    AdminOrderDetailResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            order:
              $ref: "#/components/schemas/AdminOrderSummary"
            quote:
              $ref: "#/components/schemas/Quote"
            application:
              $ref: "#/components/schemas/Application"
            payments:
              type: array
              items:
                type: object
                additionalProperties: true
            consents:
              type: array
              items:
                type: object
                additionalProperties: true
            stateTransitions:
              type: array
              items:
                type: object
                additionalProperties: true
            certificates:
              type: array
              items:
                $ref: "#/components/schemas/Certificate"
...
    CertificateResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Certificate"
    FulfilmentQueueResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              order:
                $ref: "#/components/schemas/AdminOrderSummary"
              ageHours:
                type: number
              breached:
                type: boolean
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    QuoteFunnelReportResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            quoteStarted:
              type: integer
            quoteSubmitted:
              type: integer
            quoteResultsViewed:
              type: integer
...